+++ CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - options for initApplicationCommands
 
+### Changed
+
+- `toObject` > `toJSON`
+
 ## [5.9.4] - 2021-09-22
 
 ### Added